fix warnings

+++ src/layer/riscv/rvv_mathfun.h
@@ -554,7 +554,7 @@ _RVV_FLOAT32_ERFC_OP(8, 4)
         std::vector<float> tmpy(vl);                                                                    \
         vse32_v_f32m##LMUL(tmpx.data(), a, vl);                                                         \
         vse32_v_f32m##LMUL(tmpy.data(), b, vl);                                                         \
-        for (int i = 0; i < vl; i++)                                                                    \
+        for (size_t i = 0; i < vl; i++)                                                                 \
         {                                                                                               \
             tmpx[i] = atan2(tmpx[i], tmpy[i]);                                                          \
         }                                                                                               \
